chore: decompose template breakdown scripts
