+++ wxt.config.ts
@@ -38,15 +38,21 @@ export default defineConfig({
 			{
 				resources: [
 					"css/base.css",
+					"css/income-breakdown.css",
+					"css/template-apply-breakdown.css",
 					"content-scripts/content.css",
-					"content-scripts/income-breakdown.css",
-					"content-scripts/template-apply-breakdown.css",
 					"lib/d3.min.js",
 					"lib/d3-sankey.min.js",
 					"content-scripts/income-breakdown.js",
 					"content-scripts/category-template-insights.js",
 					"content-scripts/schedule-highlight.js",
 					"content-scripts/template-apply-breakdown.js",
+					"content-scripts/template-apply-breakdown/actual-data.js",
+					"content-scripts/template-apply-breakdown/constants.js",
+					"content-scripts/template-apply-breakdown/dom.js",
+					"content-scripts/template-apply-breakdown/money.js",
+					"content-scripts/template-apply-breakdown/priority-plan.js",
+					"content-scripts/template-apply-breakdown/templates.js",
 				],
 				matches: ["<all_urls>"],
 			},
